generate-2fa/requirements.txt
psycopg2-binary
qrcode
pillow
pyotp
cryptography